% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{availablePHGMethods}
\alias{availablePHGMethods}
\alias{availablePHGMethods,BrapiCon-method}
\title{Retrieve available PHG method data from BrAPI connection}
\usage{
availablePHGMethods(object)

\S4method{availablePHGMethods}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{variantTables} endpoint of a BrAPI
  server.
}
