% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_server.R
\docType{class}
\name{BrapiCon-class}
\alias{BrapiCon-class}
\title{An S4 BrapiCon Class}
\description{
Class \code{BrapiCon} defines a \code{rPHG}
   Class for storing BrAPI connection data.
}
\section{Slots}{

\describe{
\item{\code{host}}{A URL to a BrAPI server.}

\item{\code{port}}{The host port.}

\item{\code{protocol}}{Which protocol must be used to fetch the desired data? Must
be either \code{http} or \code{https}.}

\item{\code{version}}{BrAPI version number. Must be either \code{"v1"} or
\code{"v2"}.}

\item{\code{token}}{API authorization token.}

\item{\code{url}}{BrAPI server URL.}
}}

