% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_classes.R
\docType{class}
\name{BrapiConPHG-class}
\alias{BrapiConPHG-class}
\title{An S4 BrapiConPHG Class}
\description{
Class \code{BrapiConPHG} defines a \code{rPHG}
   Class for storing BrAPI connection data plust PHG coordinate info.
}
\section{Slots}{

\describe{
\item{\code{methodID}}{A PHG method identifier.}

\item{\code{refRangeFilter}}{Reference range selection URL parameters.}

\item{\code{sampleFilter}}{Sample / taxa selection URL parameters.}
}}

