% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{brapiVersion}
\alias{brapiVersion}
\alias{brapiVersion,PHGServerCon-method}
\title{Return BrAPI version ID}
\usage{
brapiVersion(object, ...)

\S4method{brapiVersion}{PHGServerCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the version ID for a BrAPI-compliant PHG server
}
