% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_graph.R
\name{buildPHGraph}
\alias{buildPHGraph}
\title{Helper function to build PHGraph object}
\usage{
buildPHGraph(
  phgMethodObj,
  chrom = NULL,
  includeSequence = FALSE,
  includeVariants = FALSE
)
}
\arguments{
\item{phgMethodObj}{A \code{\linkS4class{PHGMethod}} object.}

\item{chrom}{A vector of chromosomes to include in graph. If NULL, defaults
to all. To specify multiple chromosome, pass as a vector of strings (i.e.
\code{c("1", "2", "3")}). Is currently only used for haplotypes.}

\item{includeSequence}{Whether to include sequences in haplotype nodes.
Is currently only used for haplotypes. NOTE: this will greatly increase
memory consumption!}

\item{includeVariants}{Whether to include variant contexts in haplotype
nodes. Is currently only used for haplotypes. NOTE: this will greatly
increase memory consumption!}
}
\description{
Creates a \code{\linkS4class{PHGraph}} object to be used to build and store
an \code{rJava} reference object pointing to a \code{HaplotypeGraph} object
from the PHG API.
}
