% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{calcMutualInfo}
\alias{calcMutualInfo}
\alias{calcMutualInfo,PHGDataSet-method}
\title{Calculate mutual information for a given PHG data set}
\usage{
calcMutualInfo(object, ...)

\S4method{calcMutualInfo}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG} dataset containing haplotype and reference
range information}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Calculates mutual information from an rPHG data set object containing
haplotype id and reference range information. Mutual information quantifies
the "amount of information" obtained about one random variable through
observing the other random variable. This will calcuate the the mutual
information across all pairs of reference ranges.
}
