% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{calls}
\alias{calls}
\alias{calls,BrapiCon-method}
\title{Retrieve calls from BrAPI connection}
\usage{
calls(object)

\S4method{calls}{BrapiCon}(object)
}
\arguments{
\item{object}{a \linkS4class{BrapiCon} object.}
}
\description{
Retrieves data from the \code{calls} endpoint of a BrAPI
  server.
}
