% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_support.R
\name{configCatcher}
\alias{configCatcher}
\title{Logic support for config files}
\usage{
configCatcher(configFile)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}
}
\description{
Provides logic checking for config files used in PHG creation.
}
\author{
Brandon Monier
}
