% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingSupport.R
\name{configFileMaker}
\alias{configFileMaker}
\title{Create PHG config file (WIP)}
\usage{
configFileMaker(
  dbName,
  dbType = c("sqlite", "postgresql"),
  fileName = NULL,
  exportPath = NULL,
  user = NULL,
  password = NULL
)
}
\arguments{
\item{dbName}{Database name (\code{*.db}).}

\item{dbType}{What type of database is this?}

\item{fileName}{Configuration file name (\code{*.txt}).}

\item{exportPath}{Specify a specific export path.}

\item{user}{Username for database.}

\item{password}{Password for database.}
}
\description{
This function will create a config file on the fly for basic
   PHG databases built via SQLite. This is a \strong{work in progress.}
   \strong{FOR ADVANCED USE ONLY.}
}
