% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_local.R
\name{configFilePath}
\alias{configFilePath}
\alias{configFilePath,PHGLocalCon-method}
\title{Return file path of configuration file}
\usage{
configFilePath(object, ...)

\S4method{configFilePath}{PHGLocalCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the file path for a configuration file to a PHG database
}
