% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createConfigFile}
\alias{createConfigFile}
\title{Create mock config file}
\usage{
createConfigFile(
  file,
  host = "localhost",
  user = "user",
  password = "sqlite",
  dbType = "sqlite",
  dbPath = NULL
)
}
\arguments{
\item{file}{User defined output file}

\item{host}{Host service for database}

\item{user}{User ID for database access}

\item{password}{Password for database access}

\item{dbType}{Database architecture used}

\item{dbPath}{P}
}
\description{
Creates a temporary PHG configuration file to access the
   provided example database. Mainly for debugging and educational
   purposes.
}
