% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_local.R
\name{dbName}
\alias{dbName}
\alias{dbName,PHGLocalCon-method}
\title{Return name of DB}
\usage{
dbName(object, ...)

\S4method{dbName}{PHGLocalCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the name for a given PHG database
}
