% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_local.R
\name{dbType}
\alias{dbType}
\alias{dbType,PHGLocalCon-method}
\title{Return type of DB}
\usage{
dbType(object, ...)

\S4method{dbType}{PHGLocalCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the type (e.g. postgres or sqlite) for a given PHG database
}
