% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{filterRefRanges}
\alias{filterRefRanges}
\title{Filter reference ranges from given PHG method}
\usage{
filterRefRanges(x, gr = NULL, chromosome = NULL, start = NULL, end = NULL)
}
\arguments{
\item{x}{A \code{BrapiConPHG} object.}

\item{gr}{A \code{GRanges} object. Houses genomic range information for
filter.}

\item{chromosome}{A vector of chromosome ids of type \code{character}. Can
be of length one to size \code{n}. If used, this will return all reference
ranges within a given chromosome.}

\item{start}{A vector of start positions of type \code{numeric}. If used,
an equal number of \code{end} elements will be needed to avoid error.}

\item{end}{A vector of end positions of type \code{numeric}. These will
link up with the \code{start} positions. Must be equal to the \code{start}
parameter.}
}
\description{
Filters reference ranges for a given PHG method by
  manipulation of BrAPI samples URL call. For a given query, reference
  ranges will be returned if they overlap with a user-defined range.
  Uses 1-based coordinate information.
}
