% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{filterSamples}
\alias{filterSamples}
\title{Filter samples from given PHG method}
\usage{
filterSamples(x, samples)
}
\arguments{
\item{x}{A \code{BrapiConPHG} object.}

\item{samples}{A vector of taxa ID of type \code{character}.}
}
\description{
Filters samples for a given PHG method by manipulation of BrAPI
  samples URL call. Returns exact matches only. If query is not exact match,
  no data will be returned for that given sample.
}
