% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flapjackExport.R
\name{flapjackExport}
\alias{flapjackExport}
\title{Export a PHG object to Flapjack file formats.}
\usage{
flapjackExport(phgObject, outputName = NULL)
}
\arguments{
\item{phgObject}{A PHG object.}

\item{outputName}{A specified output name for your Flapjack files. Defaults
to \code{NULL}. If \code{NULL} file name will be \code{phg_output}}
}
\description{
This function will take a PHG object and export specified
  ranges to a Flapjack file format. Take note that in order for output to
  be generated, you will have to build your PHG with a the parameter
  \code{includeVariant} to \code{TRUE}.
}
