% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_api.R
\name{getSequencesByHapId}
\alias{getSequencesByHapId}
\alias{getSequencesByHapId,PHGDataSet-method}
\alias{getSequencesByHapId,BrapiConPHG-method}
\title{Retrieve sequence data from rPHG data objects by haplotype ID}
\usage{
getSequencesByHapId(object, hapId)

\S4method{getSequencesByHapId}{PHGDataSet}(object, hapId)

\S4method{getSequencesByHapId}{BrapiConPHG}(object, hapId = NULL)
}
\arguments{
\item{object}{A \code{PHGDataSet} or \code{BrapiConPHG} object.}

\item{hadId}{A numeric vector of haplotype IDs.}
}
\value{
A \code{Biostrings::DNAStringSet} object
}
\description{
Returns sequence information for a user-defined set of
   haplotype IDs from either a local PHG database object
   (\code{PHGDataSet}) or a BrAPI connection pointing to a specific PHG
   method (\code{BrapiConPHG}).
}
