% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_api.R
\name{getSequencesByRefRange}
\alias{getSequencesByRefRange}
\alias{getSequencesByRefRange,PHGDataSet-method}
\alias{getSequencesByRefRange,BrapiConPHG-method}
\title{Retrieve sequence data from rPHG data objects by reference range}
\usage{
getSequencesByRefRange(object, refRanges = NULL, gRanges = NULL)

\S4method{getSequencesByRefRange}{PHGDataSet}(object, refRanges = NULL, gRanges = NULL)

\S4method{getSequencesByRefRange}{BrapiConPHG}(object, refRanges = NULL, gRanges = NULL)
}
\arguments{
\item{object}{A \code{PHGDataSet} or \code{BrapiConPHG} object.}

\item{refRanges}{A numeric vector of reference range IDs.}

\item{gRanges}{A \code{GRanges} object.}
}
\value{
A \code{Biostrings::DNAStringSet} object
}
\description{
Returns sequence information for a user-defined set of
   reference ranges from either a local PHG database object
   (\code{PHGDataSet}) or a BrAPI connection pointing to a specific PHG
   method (\code{BrapiConPHG}).
}
