% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_brapi.R
\name{getVTList}
\alias{getVTList}
\title{Retrieve variant table BrAPI URLs}
\usage{
getVTList(x)
}
\arguments{
\item{x}{A \code{BrapiConPHG} object.}
}
\description{
Returns a list of three BrAPI endpoints: (1) sample, (2)
  variants (i.e. reference ranges), and (3) table info.
}
