% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphBuilder.R
\name{graphBuilder}
\alias{graphBuilder}
\title{Test PHG builder function}
\usage{
graphBuilder(configFile, myMethods, myChrom = NULL,
  myIncludeSequence = TRUE, myIncludeVariant = FALSE)
}
\arguments{
\item{configFile}{Configuration file for DB.}

\item{myMethods}{Pairs of method calls - passed as string.}

\item{myChrom}{List of chromosomes to include in graph. If NULL, defaults
to all.}

\item{myIncludeSequence}{Whether to include sequences in haplotype nodes.
(ADVANCED)}

\item{myIncludeVariant}{Whether to include variant contexts in haplotype
nodes. (ADVANCED)}
}
\description{
R wrapper to build a PHG graph object for downstream use.
}
