% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hapIDMatrix.R
\name{hapIDMatrix}
\alias{hapIDMatrix}
\title{Generate a haplotype ID matrix}
\usage{
hapIDMatrix(phgObject)
}
\arguments{
\item{phgObject}{A PHG object.}
}
\description{
Generates a haplotype ID matrix from a PHG object.
}
\author{
Brandon Monier

Peter Bradbury
}
