% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_haplotype_graph.R
\name{HaplotypeGraph-validity}
\alias{HaplotypeGraph-validity}
\title{HaplotypeGraph validation}
\arguments{
\item{object}{A \code{HaplotypeGraph} object.}
}
\description{
Checks if \code{HaplotypeGraph} class objects are valid.
}
