% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{httProtocol}
\alias{httProtocol}
\alias{httProtocol,PHGServerCon-method}
\title{Return protocol value}
\usage{
httProtocol(object, ...)

\S4method{httProtocol}{PHGServerCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the protocol information for a given object
}
