% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{isDemo}
\alias{isDemo}
\alias{isDemo,PHGMethod-method}
\title{Return demo state}
\usage{
isDemo(object, ...)

\S4method{isDemo}{PHGMethod}(object)
}
\arguments{
\item{object}{an \code{rPHG} method object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns logical of demo state for \code{rPHG} objects
}
