% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_graph.R
\name{javaMemoryAddress}
\alias{javaMemoryAddress}
\alias{javaMemoryAddress,HaplotypeGraph-method}
\title{Return \code{rJava} reference object}
\usage{
javaMemoryAddress(object, ...)

\S4method{javaMemoryAddress}{HaplotypeGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the \code{rJava} memory reference for a given \code{rPHG} object
}
