% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_utilities.R
\name{json2igraph}
\alias{json2igraph}
\title{Parse graph data}
\usage{
json2igraph(object, dbID)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}

\item{dbID}{A PHG method.}
}
\description{
Parses graph information from JSON structures
}
