% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_utilities.R
\name{json2tibble}
\alias{json2tibble}
\title{JSON to tibble converter}
\usage{
json2tibble(object, ep)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}

\item{ep}{A specified endpoint to request.}
}
\value{
A \code{tibble} object.
}
\description{
Converts a requested JSON object to a \code{tibble}-based
  \code{data.frame} object.
}
