% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_haplotype_graph.R
\name{numberOfChromosomes}
\alias{numberOfChromosomes}
\alias{numberOfChromosomes,HaplotypeGraph-method}
\title{Return number of chromosomes}
\usage{
numberOfChromosomes(object, ...)

\S4method{numberOfChromosomes}{HaplotypeGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of chromosomes for a given object
}
