% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_graph.R
\name{numberOfNodes}
\alias{numberOfNodes}
\alias{numberOfNodes,PHGraph-method}
\title{Return number of nodes}
\usage{
numberOfNodes(object, ...)

\S4method{numberOfNodes}{PHGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of nodes for a given object
}
