% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_graph.R
\name{numberOfRefRanges}
\alias{numberOfRefRanges}
\alias{numberOfRefRanges,PHGraph-method}
\title{Return number of reference ranges}
\usage{
numberOfRefRanges(object, ...)

\S4method{numberOfRefRanges}{PHGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of reference ranges for a given object
}
