% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{numHaploPerRange}
\alias{numHaploPerRange}
\title{Get the number of haplotypes per range in physical position segment}
\usage{
numHaploPerRange(phgObject, chr = NULL, start = 0, end = NULL)
}
\arguments{
\item{phgObject}{A PHG object.}

\item{chr}{What chromosome do you want to inspect? Defaults to \code{NULL}.
If \code{NULL}, all chromsomes will be selected.}

\item{start}{Start position of chromosome. Defaults to \code{0}.}

\item{end}{End position of chromosome. Defaults to \code{NULL}. If
\code{NULL}, the whole chromosome will be analyzed.}
}
\description{
Get the number of haplotypes per range in physical position segment
}
\author{
Jean-Luc Jannink
}
