% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{numHaploPerRefRange}
\alias{numHaploPerRefRange}
\alias{numHaploPerRefRange,PHGDataSet-method}
\title{Return number of haplotypes per reference range}
\usage{
numHaploPerRefRange(object, ...)

\S4method{numHaploPerRefRange}{PHGDataSet}(object)
}
\arguments{
\item{object}{an \code{rPHG} dataset containing haplotype and reference
range information}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of unique haplotype IDs per reference range in an rPHG
dataset
}
