% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_brapi.R
\name{parseJSON}
\alias{parseJSON}
\title{URL checker}
\usage{
parseJSON(url, verbose = FALSE)
}
\arguments{
\item{url}{A BrAPI URL endpoint.}

\item{verbose}{Do you want messages shown?}
}
\description{
Checks and parses URL inputs to list data from JSON text.
}
