% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathMatrix.R
\name{pathMatrix}
\alias{pathMatrix}
\title{Generate a path matrix}
\usage{
pathMatrix(configFile, pathDir)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}

\item{pathDir}{A directory to path files.}
}
\description{
Generates a matrix of paths from a directory path of
  \code{path.txt} files
}
\author{
Brandon Monier

Peter Bradbury
}
