% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_matrix.R
\name{pathsForMethod}
\alias{pathsForMethod}
\title{Generate a matrix for all the paths for \code{pathMethod}}
\usage{
pathsForMethod(configFile, pathMethod)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}

\item{pathMethod}{The name of the path method in the PHG DB}
}
\description{
Returns a \code{matrix} object of haplotype ids with taxa name
  for row names and reference range id for the column name.
}
\author{
Brandon Monier

Peter Bradbury
}
