% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con.R
\docType{class}
\name{PHGCon-class}
\alias{PHGCon-class}
\title{A PHGCon Class}
\description{
A \code{PHGCon} class is a parent class for 
local config file data.
}
\section{Slots}{

\describe{
\item{\code{phgType}}{What type of PHG connection is this?}

\item{\code{host}}{Location path of local SQLite, Postgres, or server database}
}}

