% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con.R
\name{PHGCon-validity}
\alias{PHGCon-validity}
\title{PHGCon validation}
\arguments{
\item{object}{A \code{\linkS4class{PHGCon}} object}
}
\description{
Checks for correct data entry into \code{PHGCon} class
}
