% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{phgConObj}
\alias{phgConObj}
\alias{phgConObj,PHGMethod-method}
\title{Return a PHG connection object}
\usage{
phgConObj(object, ...)

\S4method{phgConObj}{PHGMethod}(object)
}
\arguments{
\item{object}{an \code{rPHG} method object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns an \code{rPHG} connection object
}
