% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_dataset.R
\docType{class}
\name{PHGDataSet-class}
\alias{PHGDataSet-class}
\title{A PHGDataSet class}
\description{
A class to represent practical haplotype graph data which is
  wrapped in a \code{RangedSummarizedExperiment} class.
}
