% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_local.R
\docType{class}
\name{PHGLocalCon-class}
\alias{PHGLocalCon-class}
\alias{show,PHGLocalCon-method}
\title{A PHGLocalCon Class}
\usage{
\S4method{show}{PHGLocalCon}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{PHGLocalCon}} object}
}
\description{
A \code{PHGLocalCon} class defines a \code{rPHG} class for storing
local config file data.

Prints out information regarding properties from the \code{PHGLocalCon}
class to the console
}
\section{Slots}{

\describe{
\item{\code{dbName}}{Name of database}

\item{\code{dbType}}{Type of database}

\item{\code{configFilePath}}{Path to configuration file}
}}

