% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_local.R
\name{PHGLocalCon-validity}
\alias{PHGLocalCon-validity}
\title{PHGLocalCon validation}
\arguments{
\item{object}{A \code{\linkS4class{PHGLocalCon}} object}
}
\description{
Checks for correct data entry into \code{PHGLocalCon} class
}
