% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_local.R
\name{PHGLocalCon}
\alias{PHGLocalCon}
\title{Helper functio to construct a \code{PHGLocalCon} object}
\usage{
PHGLocalCon(file)
}
\arguments{
\item{file}{A path to a PHG configuration file}
}
\description{
Creates a \code{\linkS4class{PHGLocalCon}} object to be used to read PHG
DB data for a given set of PHG-related methods.
}
