% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\docType{class}
\name{PHGMethod-class}
\alias{PHGMethod-class}
\title{A PHGMethod Class}
\description{
Class \code{PHGMethod} defines a \code{rPHG} Class for storing
a "committed" PHG method to return data against.
}
\section{Slots}{

\describe{
\item{\code{methodID}}{A PHG method identifier.}

\item{\code{phgConObj}}{A \code{\linkS4class{PHGCon}} object}
}}

