% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\name{PHGMethod-validity}
\alias{PHGMethod-validity}
\title{PHGMethod validation}
\arguments{
\item{object}{A \code{PHGMethod} object.}
}
\description{
Checks if \code{PHGMethod} class objects are valid.
}
