% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\name{PHGMethod}
\alias{PHGMethod}
\title{Helper function to construct PHGMethod object}
\usage{
PHGMethod(phgConObj, methodID)
}
\arguments{
\item{phgConObj}{A \code{\linkS4class{PHGCon}} object.}

\item{methodID}{A PHG method identifier.}
}
\description{
Creates a \code{\linkS4class{PHGMethod}} object to be used to read and
filter data from a given PHG connection object using a verified PHG method.
}
