% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_graph.R,
%   R/class_phg_method.R
\name{phgMethodId}
\alias{phgMethodId}
\alias{phgMethodId,PHGraph-method}
\alias{phgMethodId,PHGMethod-method}
\title{Return method ID}
\usage{
phgMethodId(object, ...)

\S4method{phgMethodId}{PHGraph}(object)

\S4method{phgMethodId}{PHGMethod}(object)
}
\arguments{
\item{object}{an \code{rPHG} method object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns a method ID string for a given \code{rPHG} method class
}
