% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_haplotype_graph.R
\name{phgMethodType}
\alias{phgMethodType}
\alias{phgMethodType,HaplotypeGraph-method}
\title{Return method ID type}
\usage{
phgMethodType(object, ...)

\S4method{phgMethodType}{HaplotypeGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} method object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns a method ID type for a given \code{rPHG} method class
}
