% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_graph.R
\docType{class}
\name{PHGraph-class}
\alias{PHGraph-class}
\alias{show,PHGraph-method}
\title{A PHGraph Class}
\usage{
\S4method{show}{PHGraph}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{PHGraph}} object}
}
\description{
Class \code{PHGraph} defines a \code{rPHG} Class for storing
a \code{HaplotypeGraph} object defined in the PHG API

Prints out information regarding properties from the \code{PHGraph}
class to the console
}
\section{Slots}{

\describe{
\item{\code{methodID}}{A \code{\linkS4class{PHGMethod}} object}

\item{\code{nChrom}}{Number of chromosomes}

\item{\code{nNodes}}{Number of nodes}

\item{\code{nRefRanges}}{Number of reference ranges}

\item{\code{nTaxa}}{Number of taxa}

\item{\code{jHapGraph}}{An \code{rJava} \code{jobjRef} object representing a
\code{HaplotypeGraph} class in the PHG API}

\item{\code{jMemAddress}}{An identifier string to the JVM memory space}
}}

