% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{phGraph}
\alias{phGraph}
\alias{phGraph,BrapiCon-method}
\title{Retrieve graph data from BrAPI connection}
\usage{
phGraph(object, dbID)

\S4method{phGraph}{BrapiCon}(object, dbID)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}

\item{dbID}{A PHG method.}
}
\description{
Retrieves data from the \code{graphs} endpoint of a BrAPI
  server.
}
