% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_server.R
\docType{class}
\name{PHGServerCon-class}
\alias{PHGServerCon-class}
\alias{show,PHGServerCon-method}
\title{An PHGServerCon Class}
\usage{
\S4method{show}{PHGServerCon}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{PHGServerCon}} object}
}
\description{
Class \code{PHGServerCon} defines a \code{rPHG}
   Class for storing BrAPI connection data.

Prints out information regarding properties from the \code{PHGServerCon}
class to the console
}
\section{Slots}{

\describe{
\item{\code{port}}{The host port.}

\item{\code{protocol}}{Which protocol must be used to fetch the desired data? Must
be either \code{http} or \code{https}.}

\item{\code{version}}{BrAPI version number. Must be either \code{"v1"} or
\code{"v2"}.}

\item{\code{token}}{API authorization token.}

\item{\code{url}}{BrAPI server URL.}
}}

