% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_server.R
\name{PHGServerCon-validity}
\alias{PHGServerCon-validity}
\title{BrAPI connection validation}
\arguments{
\item{object}{A \code{PHGServerCon} object.}
}
\description{
Checks if \code{PHGServerCon} class objects are valid.
}
