% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_server.R
\name{PHGServerCon}
\alias{PHGServerCon}
\title{PHGServerCon object constructor}
\usage{
PHGServerCon(
  host,
  port = NULL,
  protocol = c("https", "http"),
  version = c("v2", "v1")
)
}
\arguments{
\item{host}{A user defined URL}

\item{port}{The host port. If \code{NULL}, a default port (e.g. \code{80} or
\code{443}) will be used depending on protocol.}

\item{protocol}{Which protocol must be used to fetch the desired data? Must
be either \code{http} or \code{https}. Defaults to \code{http}.}

\item{version}{BrAPI version number. Must be either \code{"v1"} or
\code{"v2"}. Defaults to \code{v2}.}
}
\value{
A \code{PHGServerCon} object.
}
\description{
\code{PHGServerCon} is the primary container for housing BrAPI
   connection information.
}
