% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con.R
\name{phgType}
\alias{phgType}
\alias{phgType,PHGCon-method}
\title{Return type of PHG connection}
\usage{
phgType(object, ...)

\S4method{phgType}{PHGCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the PHG type for a given \code{rPHG} local or server connection object
}
