% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_plot_dot.R
\name{plotDot}
\alias{plotDot}
\title{Generate dot plot from Anchor files}
\usage{
plotDot(
  anchorPath,
  querySeqId = NULL,
  refSeqId = NULL,
  queryLab = NULL,
  refLab = NULL,
  colorId = c("strand", "score")
)
}
\arguments{
\item{anchorPath}{Path to anchor file.}

\item{querySeqId}{Vector of sequence IDs (query).}

\item{refSeqId}{Vector of sequence IDs (reference).}

\item{queryLab}{Optional label for query axis.}

\item{refLab}{Optional label for reference axis.}

\item{colorId}{How to color plots (\code{strand} or \code{score})}
}
\description{
This function will generate a dot plot (collinearity) from an Anchor
File generated from the program
[AnchorWave](https://github.com/baoxingsong/AnchorWave). This is one of the
preliminary steps in creating a PHG database.
}
