% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsAndVisualization.R
\name{plotMutualInfo}
\alias{plotMutualInfo}
\title{Calculate and plot mutual information between a set of reference ranges}
\usage{
plotMutualInfo(phgHapIDMat = NULL, phgObject = NULL,
  gameteNames = NULL, refRanges)
}
\arguments{
\item{phgHapIDMat}{The output of the \code{hapIDMatrix()} function.}

\item{phgObject}{A PHG object.}

\item{gameteNames}{Specified gamete names. If \code{NULL}, gamete names will
default to taxa IDs (haplottype ID matrix row names).}

\item{refRanges}{What reference ranges you wan to specify?}
}
\description{
Mutual information quantifies the “amount of information”
  obtained about one random variable through observing the other random
  variable.
}
