% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{plotNumHaplo}
\alias{plotNumHaplo}
\title{Plot the number of haplotypes}
\usage{
plotNumHaplo(haploData)
}
\arguments{
\item{haploData}{The output of \code{numHaploPerRange()}}
}
\description{
This function will plot the number of haplotypes. Its input
  will be the output of the \code{numHaploPerRange()} function.
}
