% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{readHaplotypeIds}
\alias{readHaplotypeIds}
\title{Return haplotype IDs}
\usage{
readHaplotypeIds(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets haplotype ID for given samples and reference ranges for PHG method
}
