% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathMatrix.R
\name{readMappingsForLineName}
\alias{readMappingsForLineName}
\title{Retrieve read mapping information from PHG database.}
\usage{
readMappingsForLineName(
  configFile,
  lineName,
  readMappingMethodName,
  haplotypeMethodName
)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}

\item{lineName}{The name of the line (taxon) for which the read mapping
information is to be retrieved. If there are multiple read mappings with
different \code{file_group_names}, they will be combined.}

\item{readMappingMethodName}{The method name for the read mappings
(only takes a single method).}

\item{haplotypeMethodName}{The haplotype method name.}
}
\description{
Return a \code{data.frame} object of read mapping information
  for a given line (i.e. taxon).
}
\author{
Brandon Monier

Peter Bradbury
}
