% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{readMappingTableInfo}
\alias{readMappingTableInfo}
\title{Retrieve read mapping records from PHG database.}
\usage{
readMappingTableInfo(configFile)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}
}
\description{
Returns an \code{S4Vectors} \code{DataFrame} object of read
  mapping record information without \code{read_mapping} data.
}
\author{
Brandon Monier

Peter Bradbury
}
