% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{readPHGDatasetFromBrapi}
\alias{readPHGDatasetFromBrapi}
\title{Read PHGDataset object from BrAPI PHG method}
\usage{
readPHGDatasetFromBrapi(object, ...)
}
\arguments{
\item{object}{A \code{BrapiConPHG} object.}

\item{...}{Additional arguments to be passed.}
}
\description{
Creates a \code{PHGDataset} object by reading sample,
  reference range, and feature data information.
}
