% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{readRefRanges}
\alias{readRefRanges}
\alias{readRefRanges,BrapiConPHG-method}
\title{Retrieve available ref range data from a given PHG method}
\usage{
readRefRanges(object)

\S4method{readRefRanges}{BrapiConPHG}(object)
}
\arguments{
\item{object}{A \code{BrapiConPHG} object.}
}
\description{
Retrieves reference range information from a given PHG method.
  Data returned is (1) chromosome, (2) start, and (3) stop coordinates.
}
