% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{readSamples}
\alias{readSamples}
\alias{readSamples,BrapiConPHG-method}
\title{Return samples IDs}
\usage{
readSamples(object, ...)

\S4method{readSamples}{BrapiConPHG}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets sample ID data for a given PHG method
}
