% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{readTable}
\alias{readTable}
\alias{readTable,BrapiConPHG-method}
\title{Retrieve available table data from a given PHG method}
\usage{
readTable(object)

\S4method{readTable}{BrapiConPHG}(object)
}
\arguments{
\item{object}{A \code{BrapiConPHG} object.}
}
\description{
Retrieves table information from a given PHG method.
  Data returned is a \code{matrix} object.
}
