% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{references}
\alias{references}
\alias{references,BrapiCon-method}
\title{Retrieve reference data from BrAPI connection}
\usage{
references(object)

\S4method{references}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{references} endpoint of a BrAPI
  server.
}
