% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refRangeTable.R
\name{refRangeTable}
\alias{refRangeTable}
\title{Generate a reference range table}
\usage{
refRangeTable(phgObject)
}
\arguments{
\item{phgObject}{A PHG object.}
}
\description{
Generates a reference range table from a PHG object.
}
\author{
Brandon Monier

Peter Bradbury
}
