% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{samples,BrapiCon-method}
\alias{samples,BrapiCon-method}
\title{Retrieve samples from BrAPI connection}
\usage{
\S4method{samples}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{samples} endpoint of a BrAPI
  server.
}
\seealso{
[calls()] for samples, [callsets()] for callsets
}
