% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{searchSimilarGametes}
\alias{searchSimilarGametes}
\title{Search for similar gamets}
\usage{
searchSimilarGametes(
  phgObject = NULL,
  refRanges,
  gameteName,
  fractionDiff = 0,
  phgHapIDMat = NULL
)
}
\arguments{
\item{phgObject}{A PHG object.}

\item{refRanges}{Specifed reference ranges.}

\item{gameteName}{A specified gamete name}

\item{fractionDiff}{The difference between haplotypes (either 0 or 1). See
description for further details.}

\item{phgHapIDMat}{The output of the \code{hapIDMatrix()} function. If
\code{NULL}, A hap ID matrix will be generated (if you have supplied a
PHG object).}
}
\description{
Search for inbred lines (gametes) that are similar to a
  specified gamete in specified reference ranges. Supply either a haplotype
  ID matrix or a phgObject from which to extract it. Specify a gamete name
  and reference ranges. The difference between haplotypes is either 0 (same)
  or 1 (different). Fraction of ranges that are different has to be lower or
  equal to fractionDiff. Ranges with unknown haplotypes (-1) do not count in
  the fraction. If all pairwise range comparisons have -1 the lines are
  considered dissimilar.
}
