% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{serverInfo}
\alias{serverInfo}
\alias{serverInfo,PHGServerCon-method}
\title{Return server information}
\usage{
serverInfo(object, ...)

\S4method{serverInfo}{PHGServerCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Get available BrAPI calls from BrAPI compliant PHG server
}
