% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_haplo_view.R
\name{shinyHaploViewer}
\alias{shinyHaploViewer}
\title{LD visualization application}
\usage{
shinyHaploViewer(phgObject)
}
\arguments{
\item{phgObject}{A PHG object.}
}
\description{
Runs an interactive visualizer for an LD object
}
