% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{showPHGMethods}
\alias{showPHGMethods}
\alias{showPHGMethods,PHGServerCon-method}
\title{Return available PHG methods}
\usage{
showPHGMethods(object, ...)

\S4method{showPHGMethods}{PHGServerCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns a collection of available PHG methods and metadata
}
