% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{studies}
\alias{studies}
\alias{studies,BrapiCon-method}
\title{Retrieve study data from BrAPI connection}
\usage{
studies(object)

\S4method{studies}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{studies} endpoint of a BrAPI
  server.
}
