% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{taxaByNode}
\alias{taxaByNode}
\alias{taxaByNode,PHGDataSet-method}
\title{Return taxa info for haplotype IDs}
\usage{
taxaByNode(object, ...)

\S4method{taxaByNode}{PHGDataSet}(object, samples = NULL, seqnames, start, end)
}
\arguments{
\item{object}{A \code{PHGDataSet} object}

\item{...}{Additional arguments, for use in specific methods}

\item{samples}{Samples/taxa to include in plot}

\item{seqnames}{A sequence (e.g. chromosome) ID}

\item{start}{Start position for ref ranges}

\item{end}{End position for ref ranges}
}
\description{
Returns taxa (e.g. sample) information for a select set of reference ranges.
Reference ranges are identified by a user defined genomic range consisting
of a sequence (e.g. chromosome) ID, and start and stop positions.
}
