% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_by_node_utilities.R
\name{tnHashMapToTibble}
\alias{tnHashMapToTibble}
\title{Convert PHG HashMap to tibble (house-keeping)}
\usage{
tnHashMapToTibble(x)
}
\arguments{
\item{x}{HashMap to R list}
}
\description{
Convert PHG HashMap to tibble (house-keeping)
}
