% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_con_server.R
\name{BrapiCon-validity}
\alias{BrapiCon-validity}
\title{BrAPI connection validation}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Checks if \code{BrapiCon} class objects are valid.
}
