% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{brapiURL}
\alias{brapiURL}
\alias{brapiURL,BrapiCon-method}
\title{Return URL path}
\usage{
brapiURL(object, ...)

\S4method{brapiURL}{BrapiCon}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object.}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the Uniform Resource Locator (URL) of a \code{BrapiCon} object.
}
