% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_getters_setters.R
\name{brapiURL}
\alias{brapiURL}
\alias{brapiURL,BrapiCon-method}
\title{The URL of a \code{BrapiCon} object}
\usage{
brapiURL(x)

\S4method{brapiURL}{BrapiCon}(x)
}
\arguments{
\item{x}{a \linkS4class{BrapiCon} object.}
}
\description{
get or set the Uniform Resource Locator (URL) of a
  \code{BrapiCon} object.
}
