% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{brapiVersion}
\alias{brapiVersion}
\title{Return BrAPI version ID}
\usage{
brapiVersion(object, ...)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the version ID for a BrAPI-compliant PHG server
}
