% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{calcMutualInfo}
\alias{calcMutualInfo}
\title{Calculate the mutual information between a set of reference ranges}
\usage{
calcMutualInfo(
  phgObject = NULL,
  refRanges,
  gameteNames = NULL,
  phgHapIDMat = NULL
)
}
\arguments{
\item{phgObject}{A PHG object.}

\item{refRanges}{What reference ranges you wan to specify?}

\item{gameteNames}{Specified gamete names. If \code{NULL}, gamete names will
default to taxa IDs (haplottype ID matrix row names).}

\item{phgHapIDMat}{The output of the \code{hapIDMatrix()} function.}
}
\description{
Mutual information quantifies the "amount of information"
  obtained about one random variable through observing the other random
  variable. Specify the gamete names over which you want to calculate and
  reference ranges.
}
