% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{callsets}
\alias{callsets}
\alias{callsets,BrapiCon-method}
\title{Retrieve callsets from BrAPI connection}
\usage{
callsets(object)

\S4method{callsets}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{calls} endpoint of a BrAPI
  server.
}
