% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphBuilder.R
\name{graphBuilder}
\alias{graphBuilder}
\title{Test PHG builder function}
\usage{
graphBuilder(configFile, methods, chrom = NULL,
  includeSequence = FALSE, includeVariant = FALSE)
}
\arguments{
\item{configFile}{Path to a configuration file for your graph database.}

\item{methods}{Pairs of method calls - passed as string.}

\item{chrom}{List of chromosomes to include in graph. If NULL, defaults
to all.}

\item{includeSequence}{Whether to include sequences in haplotype nodes.
(ADVANCED)}

\item{includeVariant}{Whether to include variant contexts in haplotype
nodes. (ADVANCED)}
}
\description{
R wrapper to build a PHG graph object for downstream use.
}
\author{
Brandon Monier

Peter Bradbury
}
