% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvcf_metrics.R
\name{gvcfMetrics}
\alias{gvcfMetrics}
\title{Generate metrics for GVCF data}
\usage{
gvcfMetrics(gvcfDir = NULL, indelReport = FALSE)
}
\arguments{
\item{gvcfDir}{Directory containing GVCF file(s) to process.}

\item{indelReport}{Do you want to generate an indel size report? Defaults
to \code{FALSE}.}
}
\description{
Generate metrics for GVCF data
}
