% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{host}
\alias{host}
\alias{host,PHGServerCon-method}
\title{Return host data}
\usage{
host(object, ...)

\S4method{host}{PHGServerCon}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the host information for a given object
}
