% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_graph.R
\name{numberOfTaxa}
\alias{numberOfTaxa}
\alias{numberOfTaxa,PHGraph-method}
\title{Return number of taxa}
\usage{
numberOfTaxa(object, ...)

\S4method{numberOfTaxa}{PHGraph}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the number of taxa for a given object
}
