% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PHGDataSet-class}
\alias{PHGDataSet-class}
\title{phgDataSet}
\description{
A class to represent a practical haplotype graph which is
  wrapped in a \code{RangedSummarizedExperiment} class.
}
