% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\name{PHGMethod}
\alias{PHGMethod}
\title{Helper function to construct BrapiConPHG object}
\usage{
PHGMethod(brapiObj, x)
}
\arguments{
\item{brapiObj}{A \code{BrapiCon} object.}

\item{x}{A PHG method identifier.}
}
\description{
Creates a \code{BrapiConPHG} object to be used to read and
  filter data from a given BrAPI endpoint given a verified PHG method.
}
