% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_and_visualization.R
\name{plotGraph}
\alias{plotGraph}
\title{Visualize Graph Data}
\usage{
plotGraph(
  x,
  samples = NULL,
  sampleHighlight = NULL,
  seqId = NULL,
  start = NULL,
  end = NULL,
  colMajor = "maroon",
  colMinor = "lightgrey",
  ...
)
}
\arguments{
\item{x}{A \code{PHGDataSet} object}

\item{samples}{Samples/taxa to include in plot}

\item{sampleHighlight}{Sample path to highlight}

\item{seqId}{A sequence (e.g. chromosome) ID}

\item{start}{Start position for ref ranges}

\item{end}{End position for ref ranges}

\item{colMajor}{Highlight path color}

\item{colMinor}{Muted path color}

\item{...}{Additional parameters to pass for ref range inclusion}
}
\description{
Generates an interactive network plot for a given set of reference ranges
and a set of taxa.
}
