% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_dataset.R
\name{plotMutualInfo}
\alias{plotMutualInfo}
\alias{plotMutualInfo,PHGDataSet-method}
\title{Plot mutual information}
\usage{
plotMutualInfo(object, ...)

\S4method{plotMutualInfo}{PHGDataSet}(object)
}
\arguments{
\item{object}{A \code{PHGDataSet} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Generates a visualization of calculated mutual information for a
given set of reference ranges
}
