% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsAndVisualization.R
\name{plotNumHaplo}
\alias{plotNumHaplo}
\title{Plot the number of haplotypes}
\usage{
plotNumHaplo(refRanges)
}
\arguments{
\item{refRanges}{The output of \code{numHaploPerRange()}}
}
\description{
This function will plot the number of haplotypes. Its input
  will be the output of the \code{numHaploPerRange()} function.
}
