% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{port}
\alias{port}
\title{Return port value}
\usage{
port(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns the port information for a given object
}
