% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{readHaplotypeIds}
\alias{readHaplotypeIds}
\alias{readHaplotypeIds,BrapiConPHG-method}
\title{Return haplotype IDs}
\usage{
readHaplotypeIds(object, ...)

\S4method{readHaplotypeIds}{BrapiConPHG}(object, numCores = NULL, transpose = TRUE)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object}

\item{...}{Additional arguments, for use in specific methods}

\item{numCores}{Number of processing cores for faster processing times.}

\item{transpose}{Do you want to transpose table?}
}
\description{
Gets haplotype ID for given samples and reference ranges for PHG method
}
