% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R,
%   R/class_haplotype_graph.R, R/class_phg_method.R
\name{readHaplotypeIds}
\alias{readHaplotypeIds}
\alias{readHaplotypeIds,HaplotypeGraph-method}
\alias{readHaplotypeIds,PHGMethod-method}
\title{Return haplotype IDs}
\usage{
readHaplotypeIds(object, ...)

\S4method{readHaplotypeIds}{HaplotypeGraph}(object)

\S4method{readHaplotypeIds}{PHGMethod}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets haplotype ID for given samples and reference ranges for PHG method
}
