% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{readPHGDataSet}
\alias{readPHGDataSet}
\alias{readPHGDataSet,BrapiConPHG-method}
\title{Return a PHGDataSet}
\usage{
readPHGDataSet(object, ...)

\S4method{readPHGDataSet}{BrapiConPHG}(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Creates a \code{\linkS4class{PHGDataSet}} for a given PHG method. This will
return all 3 primary sources of data (samples, reference ranges, and
haplotype IDs).
}
