% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{readRefRanges}
\alias{readRefRanges}
\alias{readRefRanges,BrapiConPHG-method}
\title{Return reference ranges}
\usage{
readRefRanges(object, ...)

\S4method{readRefRanges}{BrapiConPHG}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Get reference range data for a given PHG method
}
