% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_method.R
\name{readRefRanges}
\alias{readRefRanges}
\alias{readRefRanges,PHGMethod-method}
\title{Return reference ranges}
\usage{
readRefRanges(object, ...)

\S4method{readRefRanges}{PHGMethod}(object)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Get reference range data for a given PHG method
}
