% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{readSamples}
\alias{readSamples}
\alias{readSamples,BrapiConPHG-method}
\title{Retrieve available sample data from a given PHG method}
\usage{
readSamples(object)

\S4method{readSamples}{BrapiConPHG}(object)
}
\arguments{
\item{object}{A \code{BrapiConPHG} object.}
}
\description{
Retrieves sample information from a given PHG method.
  Data returned is (1) sample name, (2) sample DB ID, (3) description,
  and (4) additional information.
}
