% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R
\name{readSamples}
\alias{readSamples}
\title{Return samples IDs}
\usage{
readSamples(object, ...)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Gets sample ID data for a given PHG method
}
