% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\docType{methods}
\name{samples}
\alias{samples}
\title{Retrieve samples from BrAPI connection}
\usage{
\S4method{samples}{BrapiCon}(object)

samples(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object.}
}
\description{
Retrieves data from the \code{samples} endpoint of a BrAPI
  server.
}
