% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_methods.R
\name{serverInfo}
\alias{serverInfo}
\alias{serverInfo,BrapiCon-method}
\title{Retrieve server info data from BrAPI connection}
\usage{
serverInfo(object)

\S4method{serverInfo}{BrapiCon}(object)
}
\arguments{
\item{object}{A \code{BrapiCon} object.}
}
\description{
Retrieves data from the \code{serverinfo} endpoint of a BrAPI
  server.
}
