% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_server.R
\name{serverInfo}
\alias{serverInfo}
\alias{serverInfo,BrapiCon-method}
\title{Return server information}
\usage{
serverInfo(object, ...)

\S4method{serverInfo}{BrapiCon}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiCon}} object}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Get avaiable BrAPI calls from BrAPI compliant PHG server
}
