% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,PHGMethod-method}
\title{Show method for PHGMethod objects}
\usage{
\S4method{show}{PHGMethod}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{PHGMethod}} object.}
}
\description{
Prints out information regarding properties from the \code{PHGMethod}
class to the console
}
