% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_phg_method.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,BrapiConPHG-method}
\title{Show method for BrapiConPHG objects}
\usage{
\S4method{show}{BrapiConPHG}(object)
}
\arguments{
\item{object}{a \code{\linkS4class{BrapiConPHG}} object.}
}
\description{
Prints out the information from the BrAPI connection object
  including server status codes. See this
  \href{https://en.wikipedia.org/wiki/List_of_HTTP_status_codes}{Wikipedia link}
  for further details about what these codes mean.
}
