% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_all_generics.R, R/class_phg_con_local.R,
%   R/class_phg_con_server.R
\name{showPHGMethods}
\alias{showPHGMethods}
\alias{showPHGMethods,PHGLocalCon-method}
\alias{showPHGMethods,PHGServerCon-method}
\title{Return available PHG methods}
\usage{
showPHGMethods(object, showAdvancedMethods = FALSE, ...)

\S4method{showPHGMethods}{PHGLocalCon}(object, showAdvancedMethods)

\S4method{showPHGMethods}{PHGServerCon}(object, showAdvancedMethods)
}
\arguments{
\item{object}{an \code{rPHG} local or server connection object}

\item{showAdvancedMethods}{Do you want to return all possible method IDs
from the database? Defaults to \code{FALSE}.}

\item{...}{Additional arguments, for use in specific methods}
}
\description{
Returns a collection of available PHG methods and metadata
}
