% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPHGMethods.R
\name{showPHGMethods}
\alias{showPHGMethods}
\title{Get DB PHG methods for graph building}
\usage{
showPHGMethods(configFile)
}
\arguments{
\item{configFile}{Configuration file for graph database.}
}
\description{
Gets all available PHG methods from the graph database
  using a path parameter to the database configuration file.
}
\author{
Brandon Monier

Peter Bradbury
}
