% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_by_node.R
\name{taxaByNode}
\alias{taxaByNode}
\title{Get taxa data for selected reference ranges}
\usage{
taxaByNode(phgObj, start = NULL, end = NULL, seqnames = NULL, rrSet = NULL)
}
\arguments{
\item{phgObj}{An object of class \code{PHGDataSet}.}

\item{start}{Start position (bp) for reference range filtering.}

\item{end}{End position (bp) for reference range filtering.}

\item{seqnames}{Sequence name (e.g. chromosome ID) for reference range
filtering.}

\item{rrSet}{A collection of reference range IDs. Defaults to \code{NULL}
if specified with an integer vector, \code{start}, \code{end}, and
\code{seqnames} parameters will be ignored.}
}
\description{
Base code to get assembly information from haplotype node objects
}
